refactor(galgame): split plugin_ocr_helpers.py from __init__.py
Move 11 OCR helpers + 2 module-level constants verbatim:
- _normalize_ocr_trigger_mode
- _normalize_reader_mode
- _session_candidate_has_text
- _pending_data_source_for_reader_mode
- _after_advance_screen_refresh_needed
- _companion_after_advance_ocr_refresh_needed
- _ocr_reader_allowed_block_reason
- _ocr_tick_block_reason
- _ocr_emit_block_reason
- _apply_ocr_decision_diagnostics
- _merge_ocr_runtime_preserving_bridge_diagnostics
- _AFTER_ADVANCE_SCREEN_REFRESH_STAGES
- _OCR_BRIDGE_DIAGNOSTIC_RUNTIME_KEYS

__init__.py re-exports them so external import paths stay compatible.
